Split tag search tests for lint
